requirements-spark.txt
pyspark>=3.5.0
psycopg2-binary>=2.9.9
PyYAML>=6.0
